Create Glance_type_Fixation_Saccade
This code checks the existing numerical values of this column and creates a new column called "Glance_Type" that checks if the numerical value is between 150 ms and 700 ms, write Fixation in the column "Glance_Type", and if the numerical value is less than 150 ms, write Saccade in the column "Glance_Type", and if the numerical value more than 700 ms, write Long in the column "Glance_Type". If there is a blank value for "GLANCE_DURATION_milliseconds", just skip it.