Create MODE_GAZE_QUADRANT
I want you for each interval between "TIMESTAMP_INITIAL" and "TIMESTAMP_FINAL", to look at the column "GAZE_QUADRANT":
Step 1: In case of blank values, just skip them.
Step 2: Between each "TIMESTAMP_INITIAL" and "TIMESTAMP_FINAL": There is a range of rows of "GAZE_QUADRANT", of different values such as "1", "4", "7", etc. I want you to calculate the MODE of this range of rows of "GAZE_QUADRANT", and give me the most frequent number and return it in a new column called "MODE_GAZE_QUADRANT". 
In step 2, in the case of two values having the same number of occurrences, I want you to return the object based on the priority rule that I will give you:
[1,4,7,3,2,5,6].